session_test: remove session_test.go
