+++ tidb_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/store/mockstore"
+	"github.com/pingcap/tidb/terror"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util"
 	"github.com/pingcap/tidb/util/auth"
@@ -176,6 +177,50 @@ func (s *testMainSuite) TestSysSessionPoolGoroutineLeak(c *C) {
 	c.Assert(after-before, Less, 3)
 }
 
+func (s *testMainSuite) TestSchemaCheckerSimple(c *C) {
+	defer testleak.AfterTest(c)()
+	lease := 5 * time.Millisecond
+	validator := domain.NewSchemaValidator(lease)
+	checker := &schemaLeaseChecker{SchemaValidator: validator}
+
+	// Add some schema versions and delta table IDs.
+	ts := uint64(time.Now().UnixNano())
+	validator.Update(ts, 0, 2, []int64{1})
+	validator.Update(ts, 2, 4, []int64{2})
+
+	// checker's schema version is the same as the current schema version.
+	checker.schemaVer = 4
+	err := checker.Check(ts)
+	c.Assert(err, IsNil)
+
+	// checker's schema version is less than the current schema version, and it doesn't exist in validator's items.
+	// checker's related table ID isn't in validator's changed table IDs.
+	checker.schemaVer = 2
+	checker.relatedTableIDs = []int64{3}
+	err = checker.Check(ts)
+	c.Assert(err, IsNil)
+	// The checker's schema version isn't in validator's items.
+	checker.schemaVer = 1
+	checker.relatedTableIDs = []int64{3}
+	err = checker.Check(ts)
+	c.Assert(terror.ErrorEqual(err, domain.ErrInfoSchemaChanged), IsTrue)
+	// checker's related table ID is in validator's changed table IDs.
+	checker.relatedTableIDs = []int64{2}
+	err = checker.Check(ts)
+	c.Assert(terror.ErrorEqual(err, domain.ErrInfoSchemaChanged), IsTrue)
+
+	// validator's latest schema version is expired.
+	time.Sleep(lease + time.Microsecond)
+	checker.schemaVer = 4
+	checker.relatedTableIDs = []int64{3}
+	err = checker.Check(ts)
+	c.Assert(err, IsNil)
+	nowTS := uint64(time.Now().UnixNano())
+	// Use checker.SchemaValidator.Check instead of checker.Check here because backoff make CI slow.
+	result := checker.SchemaValidator.Check(nowTS, checker.schemaVer, checker.relatedTableIDs)
+	c.Assert(result, Equals, domain.ResultUnknown)
+}
+
 func newStore(c *C, dbPath string) kv.Storage {
 	store, err := mockstore.NewMockTikvStore()
 	c.Assert(err, IsNil)
@@ -240,13 +285,4 @@ func match(c *C, row []types.Datum, expected ...interface{}) {
 		need := fmt.Sprintf("%v", expected[i])
 		c.Assert(got, Equals, need)
 	}
-}
-
-func mustExecFailed(c *C, se Session, sql string, args ...interface{}) {
-	r, err := exec(se, sql, args...)
-	if err == nil && r != nil {
-		// sometimes we may meet error after executing first row.
-		_, err = r.Next(goctx.Background())
-	}
-	c.Assert(err, NotNil)
 }
